+++ tests/reflection/test_table_reflection.py
@@ -1,25 +1,42 @@
-import asyncio
-import functools
-import random
-import string
+from contextlib import redirect_stdout
+from io import StringIO
 
 import pytest
 
 import edgy
 from edgy.core.db.datastructures import Index
 from edgy.testclient import DatabaseTestClient
+from edgy.utils.inspect import InspectDB
 from tests.settings import DATABASE_URL
 
 pytestmark = pytest.mark.anyio
 
 database = DatabaseTestClient(DATABASE_URL)
-models = edgy.Registry(database=edgy.Database(database, force_rollback=True))
+models = edgy.Registry(database)
+second = edgy.Registry(database=edgy.Database(database, force_rollback=False))
 
+expected_result1 = """
+class Users(edgy.ReflectModel):
+    name = edgy.CharField(max_length=255, null=False)
+    title = edgy.CharField(max_length=255, null=True)
+    id = edgy.BigIntegerField(null=False, primary_key=True)
 
-def get_random_string(length):
-    letters = string.ascii_lowercase
-    result_str = "".join(random.choice(letters) for i in range(length))
-    return result_str
+    class Meta:
+        registry = registry
+        tablename = 'users'
+""".strip()
+
+expected_result2 = """
+class Hubusers(edgy.ReflectModel):
+    name = edgy.CharField(max_length=255, null=False)
+    title = edgy.CharField(max_length=255, null=True)
+    description = edgy.CharField(max_length=255, null=True)
+    id = edgy.BigIntegerField(null=False, primary_key=True)
+
+    class Meta:
+        registry = registry
+        tablename = 'hubusers'
+""".strip()
 
 
 class User(edgy.Model):
@@ -51,7 +68,7 @@ class ReflectedUser(edgy.ReflectModel):
 
     class Meta:
         tablename = "hubusers"
-        registry = models
+        registry = second
 
 
 class NewReflectedUser(edgy.ReflectModel):
@@ -60,36 +77,17 @@ class NewReflectedUser(edgy.ReflectModel):
 
     class Meta:
         tablename = "hubusers"
-        registry = models
-
-
-@pytest.fixture(autouse=True, scope="module")
-async def create_test_database():
-    async with database:
-        await models.create_all()
-        yield
-        if not database.drop:
-            await models.drop_all()
+        registry = second
 
 
 @pytest.fixture(autouse=True, scope="function")
-async def rollback_transactions():
-    async with models.database:
-        yield
-
-
-def async_adapter(wrapped_func):
-    """
-    Decorator used to run async test cases.
-    """
-
-    @functools.wraps(wrapped_func)
-    def run_sync(*args, **kwargs):
-        loop = asyncio.get_event_loop()
-        task = wrapped_func(*args, **kwargs)
-        return loop.run_until_complete(task)
-
-    return run_sync
+async def create_test_database():
+    async with models:
+        await models.create_all()
+        async with second:
+            yield
+        if not database.drop:
+            await models.drop_all()
 
 
 async def test_can_reflect_existing_table():
@@ -134,7 +132,7 @@ async def test_can_reflect_existing_table_with_not_all_fields_and_create_record(
     user = users[1]
 
     assert user.name == "Test2"
-    assert not hasattr(user, "description")
+    assert "description" not in user.__dict__
 
 
 async def test_can_reflect_and_edit_existing_table():
@@ -163,3 +161,15 @@ async def test_can_reflect_and_edit_existing_table():
 
     assert user.name == "edgy"
     assert user.description == "updated"
+
+
+async def test_create_correct_inspect_db():
+    inflected = InspectDB(str(models.database.url))
+    out = StringIO()
+    with redirect_stdout(out):
+        inflected.inspect()
+    out.seek(0)
+    generated = out.read()
+    # indexes are not sorted and appear in any order so they are removed
+    assert expected_result1 in generated
+    assert expected_result2 in generated